% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{dBP}
\alias{dBP}
\title{Density of bivariate Poisson distribution (BP)}
\usage{
dBP(x, y, m0, m1, m2, log = FALSE, max = 500)
}
\arguments{
\item{x}{a number. The first element of the data point.}

\item{y}{a number. The second element of the data point.}

\item{m0}{a positive number. The shared mean parameter.}

\item{m1}{a positive number. The second mean parameter.}

\item{m2}{a positive number. The third mean parameter.}

\item{log}{logical. If TRUE, the log-density is returned.}

\item{max}{a positive integer. If at least one of \code{x}
and \code{y} is greater than this value, the density of
zero is returned.}
}
\value{
the BP density at the data point
}
\description{
This function calculates the density of BP model
at a data point (x, y) under a set of parameters
(m0, m1, m2).
}
\examples{

dBP(2, 2, 0, 1, 1)
dBP(100, 100, 0, 1, 1)

# Sum of joint density should equal the marginal density
sum(sapply(0:100, function(s) dBP(s,2,1,2,2)))
dpois(2, 3)

# joint(BP) > pois^2 (around diagonal)
dBP(2,2,1,2,2)
dpois(2,3)^2

# joint(BP) < pois*pois (off-diagonal)
dBP(1,3,1,2,2)
dpois(1,3)*dpois(3,3)

}
