% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzinb.R
\name{idigamma}
\alias{idigamma}
\title{Inverse digamma function}
\usage{
idigamma(y)
}
\arguments{
\item{y}{a numeric vector.}
}
\description{
inverse of digamma. digamma function is the first derivative of gamma function divided by gamma function.
}
\examples{
idigamma(2)
plot(digamma, 0.1, 3)
plot(idigamma, -10.4, 0.9)

}
\author{
Hunyong Cho, Chuwen Liu, Jinyoung Park, and Di Wu
}
