% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzinbData.R
\name{rBzinbData}
\alias{rBzinbData}
\title{rBzinbData
BZINB data generation with random normal covariates where the covariates corresponding to eta 
and gamma coincide up to the smaller dimension.}
\usage{
rBzinbData(
  n = 200,
  alpha = c(1, 1, 1),
  eta1 = c(-3, 0.5, 0.5, 0.5, 0.5),
  eta2 = c(-2, 0, 1, 0, 1),
  gamma1 = c(1, 1),
  gamma2 = c(-1, 1),
  gamma3 = c(-1, 1),
  rho = 0.3
)
}
\description{
install.packages("MASS")
}
\examples{
rBzinbData()
}
